makefile: do not delete tagged docker image
Previously, the docker image was removed after the docker target
was finished. This caused the image to be rebuilt from scratch
on each run. Leaving the tagged image on disk will allow the
cache to be used for untouched layers. This vastly improves
subsequent build times.

Future work includes exploring the use of alpine as a base image.
This could drastically reduce the size of the image, which is
currently.