apdu-signature: store name on covenant.
Previously, the name was stored within the covenant items object.
This required accessing the correct covenant type within the covenant
items union to retrieve the name. This commit moves the name up one
level to the covenant object. It allows us to remove the large switch
statement in ledger-ui.c and access the name directly.